rename folder api -> store
