+++ packages/client/src/store/index.ts
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { setupListeners } from '@reduxjs/toolkit/query';
-import { rootReducer } from 'api/reducers/rootReduser';
-import { authApi } from 'api/services/auth/authApi';
-import { rtkApi } from 'api/services/gameApi/rtk';
+import { rootReducer } from 'store/reducers/rootReduser';
+import { authApi } from 'store/services/auth/authApi';
+import { rtkApi } from 'store/services/gameApi/rtk';
 
 export const store = configureStore({
 	reducer: rootReducer,
